STCC-257 Refactored svg to png conversion, moved tests [WM,SS]
